+++ Gameboy.Interfaces/IGraphicsProcessingUnit.cs
@@ -2,7 +2,7 @@ namespace Gameboy.Interfaces
 {
     public interface IGraphicsProcessingUnit
     {
-        GBMemory GetByte(ushort absoluteAddress, MemoryRegion region);
-        void SetByte(ushort absoluteAddress, byte val, MemoryRegion region);
+        GBMemory GetByte(int absoluteAddress, MemoryRegion region);
+        void SetByte(int absoluteAddress, int val, MemoryRegion region);
     }
 }
